Changed 1604 Start dates
Changed the 1604 scenarios to occur after the signing of the Treaty of
London to better model
the Dutch Rebellion and after the start of Bocskay’s revolt to better
model the end of the Fifteen Years War.
This also removes the need to make major modifications to the history
files between 1603-1604. The wars are unchanged
since last version, this is being added to the changelog to awnser "why
are the start dates changed"